requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
motor==3.7.0
pydantic==2.10.6
python-dotenv==1.0.1
bcrypt==4.3.0
pyjwt==3.10.1
requests==2.32.3
openai==1.65.2